+++ casm/project/_ClexDescription.py
@@ -1,47 +1,59 @@
+from typing import Optional
+
 from casm.tools.shared.json_io import pretty_json
 
 
 class ClexDescription:
     """Settings for a cluster expansion
 
-    Attributes
-    ----------
-
-      name: str
-        Cluster expansion name
-
-      property: str
-        Name of the property being cluster expanded
-
-      calctype: str
-        Calctype name
-
-      ref: str
-        Reference state name
-
-      bset: str
-        Basis set
-
-      eci: str
-        ECI set name
-
+    This holds identifiers for data used in a cluster expansion, as a shortcut for
+    collecting and saving data.
     """
 
     def __init__(
         self,
         name: str,
         property: str,
-        calctype: str,
-        ref: str,
-        bset: str,
-        eci: str,
+        calctype: Optional[str] = None,
+        ref: Optional[str] = None,
+        bset: Optional[str] = None,
+        fit_id: Optional[str] = None,
+        fit_index: Optional[int] = None,
+        eci: Optional[str] = None,
     ):
         self.name = name
+        """str: Cluster expansion name"""
+
         self.property = property
+        """str: Cluster expansion name"""
+
         self.calctype = calctype
+        """Optional[str]: Calctype name"""
+
         self.ref = ref
+        """Optional[str]: Reference state name"""
+
         self.bset = bset
+        """Optional[str]: Basis set id"""
+
+        self.fit_id = fit_id
+        """Optional[str]: Fit id"""
+
+        self.fit_index = fit_index
+        """Optional[int]: Fit index, if applicable.
+        
+        Use for fitting methods that result in multiple sets of coefficients, such as
+        a distribution. 
+        """
+
         self.eci = eci
+        """Optional[str]: ECI set name
+        
+        .. deprecated:: 2.0a2
+        
+            Use :attr:`fit_id` and :attr:`fit_index` instead.
+        
+        """
 
     @staticmethod
     def from_dict(data):
@@ -51,6 +63,8 @@ class ClexDescription:
             calctype=data.get("calctype"),
             ref=data.get("ref"),
             bset=data.get("bset"),
+            fit_id=data.get("fit_id"),
+            fit_index=data.get("fit_index"),
             eci=data.get("eci"),
         )
 
@@ -59,6 +73,8 @@ class ClexDescription:
             "bset": self.bset,
             "calctype": self.calctype,
             "eci": self.eci,
+            "fit_id": self.fit_id,
+            "fit_index": self.fit_index,
             "name": self.name,
             "property": self.property,
             "ref": self.ref,