+++ casm/project/_DirectoryStructure.py
@@ -1,11 +1,12 @@
 import pathlib
+import warnings
 from typing import Optional, Union
 
 from ._ClexDescription import ClexDescription
 
 
-class DirectoryStructure:
-    """Standard CASM project directory structure
+class DirectoryStructureV1:
+    """Standard CASM v1 project directory structure
 
     This class helps constructs the standard paths specified in the CASM
     `Project directory structure reference <https://prisms-center.github.io/CASMcode_docs/formats/project_directory_structure/>`_.
@@ -32,23 +33,13 @@ class DirectoryStructure:
             )
         self.__casm_dir = ".casm"
         self.__casmdb_dir = "jsonDB"
-        self.__enum_dir = "enumerations"
         self.__bset_dir = "basis_sets"
         self.__calc_dir = "training_data"
-        self.__calculation_settings_dir = "calculation_settings"
-        self.__import_dir = "imports"
-        self.__fit_dir = "fits"
         self.__set_dir = "settings"
         self.__sym_dir = "symmetry"
         self.__clex_dir = "cluster_expansions"
-        self.__system_dir = "systems"
 
     # ** Query filesystem **
-
-    def all_enum(self):
-        """Check filesystem directory structure and return list of all enumeration
-        names"""
-        return self.__all_settings("enum", self.path / self.__enum_dir)
 
     def all_bset(self):
         """Check filesystem directory structure and return list of all basis set
@@ -59,13 +50,6 @@ class DirectoryStructure:
         """Check filesystem directory structure and return list of all calctype names"""
         return self.__all_settings(
             "calctype", self.path / self.__calc_dir / self.__set_dir
-        )
-
-    def all_calctype_v2(self):
-        """Check filesystem directory structure and return list of all calctype names
-        (v2.0)"""
-        return self.__all_settings(
-            "calctype", self.path / self.__calculation_settings_dir
         )
 
     def all_ref(self, calctype: str):
@@ -89,15 +73,6 @@ class DirectoryStructure:
             / self.__ref(ref)
             / self.__bset(bset),
         )
-
-    def all_import(self):
-        return self.__all_settings("import", self.__import_dir)
-
-    def all_fit(self):
-        return self.__all_settings("fit", self.__fit_dir)
-
-    def all_system(self):
-        return self.__all_settings("system", self.__system_dir)
 
     # ** File and Directory paths **
 
@@ -157,22 +132,6 @@ class DirectoryStructure:
             filename = f"config_props.{name}.json"
         return self.casm_dbdir() / self.__calctype(calctype) / filename
 
-    def event_list(self):
-        """Return master occ_event_list.json file path"""
-        return self.casm_dbdir() / "event_list.json"
-
-    def path_list(self):
-        """Return master path_list.json file path"""
-        return self.casm_dbdir() / "path_list.json"
-
-    def path_props(self, calctype, name: Optional[str] = None):
-        """Return path_props.json file path for given calculation type"""
-        if name is None:
-            filename = "path_props.json"
-        else:
-            filename = f"path_props.{name}.json"
-        return self.casm_dbdir() / self.__calctype(calctype) / filename
-
     def master_selection(self, otype):
         """Return location of MASTER selection file
 
@@ -186,10 +145,6 @@ class DirectoryStructure:
             return querydir / "Configuration" / "master_selection"
         elif otype == "scel":
             return querydir / "Supercell" / "master_selection"
-        elif otype == "event":
-            return querydir / "Event" / "master_selection"
-        elif otype == "path":
-            return querydir / "Path" / "master_selection"
         else:
             raise Exception("Unsupported type: " + str(otype))
 
@@ -337,7 +292,7 @@ class DirectoryStructure:
 
     def config_json(self, configname: str, calc_subdir: str = ""):
         """Return path to structure.json file"""
-        return self.configuration_dir(configname, calc_subdir) / "structure.json"
+        return self.configuration_dir(configname, calc_subdir) / "config.json"
 
     def structure_json(self, configname: str, calc_subdir: str = ""):
         """Return path to structure.json file"""
@@ -352,7 +307,7 @@ class DirectoryStructure:
             clex.calctype
         )
 
-    # -- calc_settings_dir - v1 --------
+    # -- calc_settings_dir --------
 
     def calc_settings_dir(self, clex: ClexDescription):
         """Return calculation settings directory path, for global settings from clex"""
@@ -413,11 +368,418 @@ class DirectoryStructure:
         """Return calculation reference settings directory path, for global settings"""
         return self.calc_settings_dir(clex.calctype) / self.__ref(clex.ref)
 
+    # -- Composition axes --------
+
+    def composition_axes(self):
+        """Return composition axes file path (deprecated)"""
+        return self.casm_dir() / "composition_axes.json"
+
+    def chemical_reference(self, clex: ClexDescription):
+        """Return chemical reference file path"""
+        return self.ref_dir(clex) / "chemical_reference.json"
+
+    # -- Cluster expansions --------
+
+    def property_dir(self, clex: ClexDescription):
+        """Returns path to eci directory"""
+        return self.path / self.__clex_dir / self.__clex_name(clex.property)
+
+    def eci_dir(self, clex: ClexDescription):
+        """
+        Returns path to eci directory
+
+        Parameters
+        ----------
+        clex: a casm.project.ClexDescription instance
+            Specifies the cluster expansion to get the eci directory for
+
+        Returns
+        -------
+        p: str
+            Path to the eci directory
+        """
+        return (
+            self.property_dir(clex)
+            / self.__calctype(clex.calctype)
+            / self.__ref(clex.ref)
+            / self.__bset(clex.bset)
+            / self.__eci(clex.eci)
+        )
+
+    def eci(self, clex: ClexDescription):
+        """
+        Returns path to eci.json
+
+        Parameters
+        ----------
+        clex: a casm.project.ClexDescription instance
+            Specifies the cluster expansion to get the eci.json for
+
+        Returns
+        -------
+        p: str
+            Path to the eci directory
+        """
+        return self.eci_dir(clex) / "eci.json"
+
+    # private:
+
+    def __bset(self, bset: str):
+        return "bset." + bset
+
+    def __calctype(self, calctype: str):
+        return "calctype." + calctype
+
+    def __ref(self, ref: str):
+        return "ref." + ref
+
+    def __clex_name(self, clex_name: str):
+        return "clex." + clex_name
+
+    def __eci(self, eci: str):
+        return "eci." + eci
+
+    def __all_settings(self, pattern: str, location: pathlib.Path):
+        """
+        Find all directories at 'location' that match 'pattern.something'
+        and return a std::vector of the 'something'
+        """
+
+        all = []
+        pattern += "."
+
+        # get all
+        if not location.exists():
+            return all
+
+        for child in location.iterdir():
+            if child.is_dir() and child.name[: len(pattern)] == pattern:
+                all.append(child.name[len(pattern) :])
+        return sorted(all)
+
+
+class DirectoryStructure:
+    """Standard CASM v2 project directory structure
+
+    This class helps constructs the standard paths specified in the CASM
+    `Project directory structure reference <https://prisms-center.github.io/CASMcode_docs/formats/project_directory_structure/>`_.
+
+    """
+
+    def __init__(self, path: Union[str, pathlib.Path]):
+        """
+        .. rubric:: Constructor
+
+        Parameters
+        ----------
+        path: Union[str, pathlib.Path]
+            Path to CASM project directory.
+
+        """
+        self.path = pathlib.Path(path)
+        """str: Path to CASM project."""
+
+        if self.path is None:
+            raise Exception(
+                "Error in casm.project.DirectoryStructure: "
+                f"No CASM project found containing {path}"
+            )
+        self.__casm_dir = ".casm"
+        self.__enum_dir = "enumerations"
+        self.__bset_dir = "basis_sets"
+        self.__calculation_settings_dir = "calculation_settings"
+        self.__import_dir = "imports"
+        self.__fit_dir = "fits"
+        self.__sym_dir = "symmetry"
+        self.__system_dir = "systems"
+
+    # ** Query filesystem **
+
+    def all_enum(self):
+        """Check filesystem directory structure and return list of all enumeration
+        names"""
+        return self.__all_settings("enum", self.path / self.__enum_dir)
+
+    def all_bset(self):
+        """Check filesystem directory structure and return list of all basis set
+        names"""
+        return self.__all_settings("bset", self.path / self.__bset_dir)
+
+    def all_calctype(self):
+        """Check filesystem and return list of all calctype names (v2)"""
+        return self.__all_settings(
+            "calctype", self.path / self.__calculation_settings_dir
+        )
+
+    def all_calctype_v2(self):
+        """Check filesystem directory structure and return list of all calctype names
+        (deprecated)
+
+        .. deprecated:: 2.0a2
+
+            Use :func:`~casm.project.DirectoryStructure.all_calctype` instead.
+            Will be removed in 2.0.0.
+
+        """
+        warnings.warn(
+            "The method 'casm.project.DirectoryStructure.all_calctype_v2' is "
+            "deprecated and will be removed in a future version. Use "
+            "'casm.project.DirectoryStructure.all_calctype' instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return self.__all_settings(
+            "calctype", self.path / self.__calculation_settings_dir
+        )
+
+    # TODO: update this
+    # def all_ref(self, calctype: str):
+    #     """Check filesystem directory structure and return list of all ref names for
+    #     a given calctype"""
+    #     return self.__all_settings("ref", self.calc_settings_dir(calctype))
+
+    def all_import(self):
+        return self.__all_settings("import", self.__import_dir)
+
+    def all_fit(self):
+        return self.__all_settings("fit", self.__fit_dir)
+
+    def all_system(self):
+        return self.__all_settings("system", self.__system_dir)
+
+    # ** File and Directory paths **
+
+    # -- Project directory --------
+
+    def root_dir(self):
+        """Return casm project directory path"""
+        return self.path
+
+    def prim(self):
+        """Return prim.json path"""
+        return self.casm_dir() / "prim.json"
+
+    # -- Hidden .casm directory --------
+
+    def casm_dir(self):
+        """Return hidden .casm dir path"""
+        return self.path / self.__casm_dir
+
+    def project_settings(self):
+        """Return project_settings.json path"""
+        return self.casm_dir() / "project_settings.json"
+
+    # -- Symmetry --------
+
+    def symmetry_dir(self):
+        """Return symmetry directory path"""
+        return self.path / self.__sym_dir
+
+    def lattice_point_group(self):
+        """Return lattice_point_group.json path"""
+        return self.symmetry_dir() / "lattice_point_group.json"
+
+    def factor_group(self):
+        """Return factor_group.json path"""
+        return self.symmetry_dir() / "factor_group.json"
+
+    def crystal_point_group(self):
+        """Return crystal_point_group.json path"""
+        return self.symmetry_dir() / "crystal_point_group.json"
+
+    # -- Basis sets --------
+
+    def _get_bset(
+        self,
+        clex: Optional[ClexDescription] = None,
+        bset: Optional[str] = None,
+    ):
+        if bset is None:
+            if clex is None:
+                raise Exception("One of clex, bset is required")
+            bset = clex.bset
+        return bset
+
+    def bset_dir(
+        self,
+        clex: Optional[ClexDescription] = None,
+        bset: Optional[str] = None,
+    ):
+        """Return path to directory contain basis set info"""
+        bset = self._get_bset(clex=clex, bset=bset)
+        return self.path / self.__bset_dir / self.__bset(bset=bset)
+
+    def bspecs(
+        self,
+        clex: Optional[ClexDescription] = None,
+        bset: Optional[str] = None,
+    ):
+        """Return basis function specs (bspecs.json) file path"""
+        return self.bset_dir(clex=clex, bset=bset) / "bspecs.json"
+
+    def basis(
+        self,
+        clex: Optional[ClexDescription] = None,
+        bset: Optional[str] = None,
+    ):
+        """Returns path to the basis.json file"""
+        return self.bset_dir(clex=clex, bset=bset) / "basis.json"
+
+    def clexulator_dir(
+        self,
+        clex: Optional[ClexDescription] = None,
+        bset: Optional[str] = None,
+    ):
+        """Returns path to directory containing global clexulator"""
+        bset = self._get_bset(clex=clex, bset=bset)
+        return self.bset_dir(bset=bset)
+
+    def clexulator_src(
+        self,
+        projectname: str,
+        clex: Optional[ClexDescription] = None,
+        bset: Optional[str] = None,
+        i_equiv: Optional[int] = None,
+    ):
+        """Returns path to global clexulator source file"""
+        bset = self._get_bset(clex=clex, bset=bset)
+        if i_equiv is None:
+            return self.bset_dir(bset=bset) / (projectname + f"_Clexulator_{bset}.cc")
+        else:
+            return (
+                self.bset_dir(bset=bset)
+                / str(i_equiv)
+                / (projectname + f"_Clexulator_{bset}.cc")
+            )
+
+    def clexulator_o(
+        self,
+        projectname: str,
+        clex: Optional[ClexDescription] = None,
+        bset: Optional[str] = None,
+        i_equiv: Optional[int] = None,
+    ):
+        """Returns path to global clexulator.o file"""
+        bset = self._get_bset(clex=clex, bset=bset)
+        if i_equiv is None:
+            return self.bset_dir(bset=bset) / (projectname + f"_Clexulator_{bset}.o")
+        else:
+            return (
+                self.bset_dir(bset=bset)
+                / str(i_equiv)
+                / (projectname + f"_Clexulator_{bset}.o")
+            )
+
+    def clexulator_so(
+        self,
+        projectname: str,
+        clex: Optional[ClexDescription] = None,
+        bset: Optional[str] = None,
+        i_equiv: Optional[int] = None,
+    ):
+        """Returns path to global clexulator so file"""
+        bset = self._get_bset(clex=clex, bset=bset)
+        if i_equiv is None:
+            return self.bset_dir(bset=bset) / (projectname + f"_Clexulator_{bset}.so")
+        else:
+            return (
+                self.bset_dir(bset=bset)
+                / str(i_equiv)
+                / (projectname + f"_Clexulator_{bset}.so")
+            )
+
+    def equivalents_info(
+        self,
+        clex: Optional[ClexDescription] = None,
+        bset: Optional[str] = None,
+    ):
+        """Return cluster functions (cluster_functions.json.gz) file path"""
+        return self.bset_dir(clex=clex, bset=bset) / "equivalents_info.json"
+
+    def cluster_functions(
+        self,
+        clex: Optional[ClexDescription] = None,
+        bset: Optional[str] = None,
+    ):
+        """Return cluster functions (cluster_functions.json.gz) file path"""
+        return self.bset_dir(clex=clex, bset=bset) / "cluster_functions.json.gz"
+
+    def clexulator_variables(
+        self,
+        clex: Optional[ClexDescription] = None,
+        bset: Optional[str] = None,
+    ):
+        """Return clexulator variables (variables.json.gz) file path"""
+        return self.bset_dir(clex=clex, bset=bset) / "variables.json.gz"
+
     # -- calc_settings_dir - v2 --------
 
-    def calctype_settings_dir_v2(self, calctype: str):
-        """Return global calculation settings directory path (new v2.0)"""
+    def _get_calctype(
+        self,
+        clex: Optional[ClexDescription] = None,
+        calctype: Optional[str] = None,
+    ):
+        if calctype is None:
+            if clex is None:
+                raise Exception("One of clex, calctype is required")
+            calctype = clex.calctype
+        return calctype
+
+    def calctype_settings_dir(
+        self,
+        clex: Optional[ClexDescription] = None,
+        calctype: Optional[str] = None,
+    ):
+        """Return global calculation settings directory path (v2)"""
         return self.path / self.__calculation_settings_dir / self.__calctype(calctype)
+
+    def calctype_settings_dir_v2(
+        self,
+        clex: Optional[ClexDescription] = None,
+        calctype: Optional[str] = None,
+    ):
+        """Return global calculation settings directory path (deprecated)
+
+        .. deprecated:: 2.0a2
+
+            Use :func:`~casm.project.DirectoryStructure.calctype_settings_dir` instead.
+            Will be removed in 2.0.0.
+
+
+        """
+        warnings.warn(
+            "The method 'casm.project.DirectoryStructure.calctype_settings_dir_v2' is "
+            "deprecated and will be removed in a future version. Use "
+            "'casm.project.DirectoryStructure.calctype_settings_dir' instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        calctype = self._get_calctype(clex=clex, calctype=calctype)
+        return self.path / self.__calculation_settings_dir / self.__calctype(calctype)
+
+    # -- chemical reference - v2 -------
+
+    def _get_ref(
+        self,
+        clex: Optional[ClexDescription] = None,
+        ref: Optional[str] = None,
+    ):
+        if ref is None:
+            if clex is None:
+                raise Exception("One of clex, ref is required")
+            ref = clex.ref
+        return ref
+
+    def ref_dir(self, clex: ClexDescription, calctype: str, ref: str):
+        """Return calculation reference settings directory path, for global settings"""
+        ref = self._get_ref(clex=clex, ref=ref)
+        return self.calctype_settings_dir(clex=clex, calctype=calctype) / self.__ref(
+            ref
+        )
+
+    def chemical_reference(self, clex: ClexDescription, calctype: str, ref: str):
+        """Return chemical reference file path"""
+        return self.ref_dir(clex=clex, calctype=calctype) / "chemical_reference.json"
 
     # -- Enumerations --------
 
@@ -495,7 +857,24 @@ class DirectoryStructure:
     # -- Composition axes --------
 
     def composition_axes(self):
-        """Return composition axes file path (deprecated v2.0a1)"""
+        """Return composition axes file path (deprecated)
+
+        .. deprecated:: 2.0a1
+
+            Included for v1 project compatibility. Use
+            :func:`~casm.project.DirectoryStructure.chemical_composition_axes` instead.
+            Will be removed in 2.0.0.
+
+        """
+        warnings.warn(
+            "The method 'casm.project.DirectoryStructure.composition_axes' is "
+            "deprecated and will be removed in a future version. Use "
+            "'casm.project.DirectoryStructureV1.composition_axes', "
+            "'casm.project.DirectoryStructure.chemical_composition_axes', or "
+            "'casm.project.DirectoryStructure.occupant_composition_axes' instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
         return self.casm_dir() / "composition_axes.json"
 
     def chemical_composition_axes(self):
@@ -518,55 +897,7 @@ class DirectoryStructure:
         """
         return self.casm_dir() / "occupant_composition_axes.json"
 
-    def chemical_reference(self, clex: ClexDescription):
-        """Return chemical reference file path"""
-        return self.ref_dir(clex) / "chemical_reference.json"
-
-    # -- Cluster expansions - v1 --------
-
-    def property_dir(self, clex: ClexDescription):
-        """Returns path to eci directory"""
-        return self.path / self.__clex_dir / self.__clex_name(clex.property)
-
-    def eci_dir(self, clex: ClexDescription):
-        """
-        Returns path to eci directory
-
-        Parameters
-        ----------
-        clex: a casm.project.ClexDescription instance
-            Specifies the cluster expansion to get the eci directory for
-
-        Returns
-        -------
-        p: str
-            Path to the eci directory
-        """
-        return (
-            self.property_dir(clex)
-            / self.__calctype(clex.calctype)
-            / self.__ref(clex.ref)
-            / self.__bset(clex.bset)
-            / self.__eci(clex.eci)
-        )
-
-    def eci(self, clex: ClexDescription):
-        """
-        Returns path to eci.json
-
-        Parameters
-        ----------
-        clex: a casm.project.ClexDescription instance
-            Specifies the cluster expansion to get the eci.json for
-
-        Returns
-        -------
-        p: str
-            Path to the eci directory
-        """
-        return self.eci_dir(clex) / "eci.json"
-
-    # -- Imports - v2 --------
+    # -- Imports --------
 
     def import_dir(self, id: str):
         """Return path to directory contain structure import info"""
@@ -576,13 +907,13 @@ class DirectoryStructure:
         """Return path to the file contain structure import settings"""
         return self.path / self.__import_dir / self.__import(id) / "settings.json"
 
-    # -- Fits - v2 --------
+    # -- Fits --------
 
     def fit_dir(self, fit: str):
         """Return path to directory containing fitting data"""
         return self.path / self.__fit_dir / self.__fit(fit)
 
-    # -- Systems - v2 --------
+    # -- Systems --------
 
     def system_dir(self, system: str, index: Optional[int] = None):
         """Return path to directory containing system info
@@ -731,12 +1062,6 @@ class DirectoryStructure:
 
     def __ref(self, ref: str):
         return "ref." + ref
-
-    def __clex_name(self, clex_name: str):
-        return "clex." + clex_name
-
-    def __eci(self, eci: str):
-        return "eci." + eci
 
     def __import(self, import_id: str):
         return "import." + import_id