+++ doc/conf.py
@@ -97,6 +97,7 @@ extensions = [
     "sphinxcontrib.bibtex",
     "sphinx.ext.intersphinx",
     "numpydoc",
+    "sphinx_copybutton",
 ]
 
 bibtex_bibfiles = ["refs.bib"]
@@ -182,20 +183,6 @@ html_theme_options = {
             "type": "fontawesome",
         },
     ],
-    # "favicons": [
-    #     {
-    #         "rel": "icon",
-    #         "sizes": "32x32",
-    #         "href": "favicon-32x32.png",
-    #     },
-    #     {
-    #         "rel": "icon",
-    #         "sizes": "16x16",
-    #         "href": "favicon-16x16.png",
-    #     },
-    #     {"rel": "apple-touch-icon", "sizes": "180x180", "href": "apple-touch-icon.png"},
-    # ],
-    # "primary_sidebar_end": ["primary_sidebar_end"]
 }
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -205,6 +192,8 @@ html_css_files = [
     "css/custom.css",
 ]
 html_favicon = "_static/favicon-16x16.png"
+
+copybutton_selector = "div.highlight pre, div.highlight.casm-code-block code"
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.